feat(refactor): move classes package
Move each classes package to model, controller, view package.
