+++ src/main/java/cc/oakk/racing/model/converter/Converter.java
@@ -1,4 +1,4 @@
-package cc.oakk.racing.converter;
+package cc.oakk.racing.model.converter;
 
 public interface Converter<T, E> {
     E convert(T from);
