+++ src/main/java/cc/oakk/racing/view/ResultView.java
@@ -1,8 +1,8 @@
 package cc.oakk.racing.view;
 
-import cc.oakk.racing.Car;
-import cc.oakk.racing.Ranking;
-import cc.oakk.racing.printer.Printable;
+import cc.oakk.racing.model.Car;
+import cc.oakk.racing.model.Ranking;
+import cc.oakk.racing.view.printer.Printable;
 
 public interface ResultView {
     void printResultHeader();
