+++ src/main/java/cc/oakk/racing/view/printer/Printable.java
@@ -1,4 +1,4 @@
-package cc.oakk.racing.printer;
+package cc.oakk.racing.view.printer;
 
 public interface Printable<T> {
     void print(Printer<T> printer);
