+++ src/main/java/cc/oakk/racing/view/printer/Printer.java
@@ -1,4 +1,4 @@
-package cc.oakk.racing.printer;
+package cc.oakk.racing.view.printer;
 
 public interface Printer<T> {
     void print(T source);
