+++ src/main/java/cc/oakk/racing/view/printer/RankingPrinter.java
@@ -1,7 +1,7 @@
-package cc.oakk.racing.printer;
+package cc.oakk.racing.view.printer;
 
-import cc.oakk.racing.Car;
-import cc.oakk.racing.Ranking;
+import cc.oakk.racing.model.Car;
+import cc.oakk.racing.model.Ranking;
 
 import java.util.List;
 import java.util.stream.Collectors;
